cloud_functions/cqc_fetcher/requirements.txt
functions-framework>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.0.0
google-cloud-bigquery>=3.0.0
requests>=2.25.0